requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
openai==0.28.1
numpy==1.26.1
pandas==2.1.2
python-dotenv~=1.0.0